+++ src/media/audio/pcm/mod.rs
@@ -7,14 +7,14 @@
 mod tests;
 
 mod buffer; // PcmBuffer
-// mod format; // TODO PcmRaw, PcmWav, …
+mod format; // PcmWav*
 mod spec; // PcmSpec, PcmSample
 
 crate::structural_mods! { // _mods
     _mods {
         pub use super::{
             buffer::*,
-            // format::*,
+            format::*,
             spec::*,
         };
     }